requirements.txt
requests==2.28.2
beautifulsoup4==4.11.1
psycopg2-binary==2.9.5
boto3
s3fs
awscli

